runtime: document optional now int64 args/results more, update js/wasm beforeIdle
A good number of the "now int64" arguments must be non-zero and the
results always non-zero, and a good number mean optional. The docs
weren't entirely accurate.

We might want an new "optionalTime" type for doc clarity (with a
method to get+lazy init the current time) but for now, just update
some docs.

And update the beforeIdle func in lock_js.js where it seems like a
zero int64 now value could've made its way in, which beforeIdle didn't
handle.

Change-Id: Ib28b74c1740ad8ddc777fc858258ac3434e157cd