cmd/trace: temporarily strip system goroutines for tests
If a goroutine is started within a user region, internal/trace assigns
the child goroutine a nameless region for its entire lifetime which is
assosciated the same task as the parent's region.

If the first GC occurs during a region, the GC worker goroutines will
also inherit a child region, which breaks our test expectations. Detect
these in the test and strip them.

A better fix would be to avoid including these in the analysis in the
first place (CL 416858), but that is somewhat risky late in the release
cycle.

For #53784.

Change-Id: Ic49b2e30e994369fb537bafe9f6d2fc50db009a6