runtime: skip slow smhasher tests with -quick
This skips all of the TestSmhasher tests that take >5s on my machine.
This also happens to exactly match the set of tests that is already
skipped on wasm for being too slow.

This CL reduces go test runtime -quick from ~110s to ~50s on my machine.

These are the slowest tests in -quick (see issue for list), and are
particularly good candidates to skip because -quick is used by cmd/dist
for concurrency stress testing and these tests are not concurrent.

For #53818.

Change-Id: I52ca062060770b81aa01b5494993eeb1beed0ddb