runtime: add mayAcquire annotation for trace.lock
DO NOT SUBMIT. This causes lock checking failures. See #53979.

We're missing lock edges to trace.lock that happen only rarely. Any
trace event can potentially fill up a trace buffer and acquire
trace.lock in order to flush the buffer, but this happens relatively
rarely, so we simply haven't seen some of these lock edges that could
happen.

Add a mayAcquire annotation to trace events to capture the possibility
of acquiring trace.lock even if dynamically it turns out a particular
event doesn't need to flush and acquire this lock.

With this change, we promote "fin, notifyList < traceStackTab" to
"fin, notifyList < trace" and now everything that emits trace events
with a P enters the tracer lock ranks via "trace".

This was found by inspecting the rank graph for things that didn't
make sense.

For #53789.

Change-Id: I13e2f6cd1b621cca4bed0cc13ef12e64d05c89a7