[release-branch.go1.17] bootstrap.bash: use git describe to get nicer file name
This lets the generated bootstrap files be named for tags
instead of or in addition to hashes,
which is easier to skim in directory listings.

Making the change here on release-branch.go1.17
because this is the branch we'll be using for bootstrap
toolchains after #44505 is done.

For example, with an earlier version of this commit in the tree,

	GOOS=dragonfly GOARCH=amd64 BOOTSTRAP_FORMAT=mintgz ./bootstrap.bash

builds

	gobootstrap-dragonfly-amd64-go1.17.12-1-g01c2885d70.tar.gz

instead of

	gobootstrap-dragonfly-amd64-g01c2885d70.tar.gz

Change-Id: Ifc00cd1de769bf807a0f6df643897c2f2339a073