cmd/internal/obj/arm64: mark branch instructions in optab
Currently, we judge whether we need to fix up the branch instruction
based on Optab.type_ field, but the type_ field in optab may change.
This CL marks the branch instruction in optab, and checks whether to
do fixing up according to the mark. Depending on the constant parameter
range of the branch instruction, there are three labels, BRANCH14BITS,
BRANCH19BITS and BRANCH26BITS.

Besides this CL removes the unnecessary alignment of the DWORD
instruction.

Change-Id: I993902b3fb5ad8e081dd6c441e86bcf581031835