crypto/internal/boring/bcache: make the Cache generic
Address a TODO in the comments to make bcache.Cache a generic
type. Cache is instantiated with a generic type V, where *V is
the type of the cached value. The Put() method now accepts a *V,
instead of an unsafe.Pointer which required a conversion from *V.
Likewise, the Get() method now returns a *V.

Also change atomically accessed field types to atomic.Pointer[T]
for simpler atomic APIs, and to force their use.

Change-Id: Ia8ecbc1974c5ca23f79d128072b90442f595935e