cmd/internal/obj/arm64: new path for adding new instructions
This CL adds a new path for adding new instructions, the purpose is to
simplify the effort of adding new instructions.

The new path is based on the assumption that a Go assembly instruction
(a Prog instance) corresponds to only one arm64 machine instruction.
Based on this assumption, matching a Prog to a machine instruction is
much simpler.

This CL preserves the original path, and theoretically based on the
above assumptions, all new instructions can be added with the new path.
At the same time, the old path can be used in cases where the new path
doesn't fit, such as Go’s custom assembly instructions and Go assembly
instructions correspond to multiple machine instructions.

There are three main changes:
1. The process of matching machine instructions for a Prog.
  1.1, For a Prog, first use the existing method to find a match in
       optab, if no match, look for it in the new path.
  1.2, Get all candidates in the machine instruction table through
       opcode.
  1.3, Check whether the number of arguments is the same, and whether
      the class of each argument matches.
It should be noted that multiple matches may be obtained through this
step, and ultimately we determine the final unique match through
heuristic encoding.

2, The instruction encoding process.
The new path encodes an instruction by the following formula:
Binary(Prog) = skeleton|Enc(opcode)|Enc(operand1)|Enc(operand2)|...
Each operand may have multiple elements, so
Enc(operand) = Enc(elm1)|Enc(elm2)...
So we changed the encoding of an instruction into the encoding of
operand elements, and this part of work needs to be done manually.
Skeleton is the known bits of the machine instruction.

3, The new machine instruction table.
The table was auto-generated by parsing the public Arm ISA document.
The parser is in the golang.org/x/arch repo. The instruction structure
is as follows:
type Inst struct {
	GoOp     obj.As      // Go opcode mnemonic
	ArmOp    A64Type     // Arm opcode mnemonic
	Feature  uint16      // such as "FEAT_LSE", "FEAT_CSSC"
	Skeleton uint32      // known bits
	Mask     uint32      // mask for disassembly
	Alias    bool        // whether it is an alias
	Args     []Operand   // operands, in Go order
}
The structure of operand is as follow:
type Operand struct {
	Class AClass     // operand class, register, constant etc.
	Elms []ElmType   // the elements that this operand includes
}
Note that the order of operands is consistent with the syntax order of
Go assembly instructions, so that we do not need to adjust the operand
order when encoding.

Updates #44734

Change-Id: I188814700506d76a353f398b76204667d93078d5