cmd/internal/obj/arm64: new path for adding new instructions
This CL adds a new path for adding new instructions, the purpose is to
simplify the effort of adding new instructions.

The new path is based on the assumption that a Go assembly instruction
(a Prog instance) corresponds to only one arm64 machine instruction.
Based on this assumption, matching a Prog to a machine instruction is
much simpler.

This CL preserves the original path, and theoretically based on the above
assumptions, all new instructions can be added with the new path. At the
same time, the old path can be used in the case where one Go assembly
instruction corresponds to multiple machine instructions, if such a
situation exists.

There are three main changes:
1. The process of matching machine instructions for a Prog.
1.1, For a Prog, first use the existing method to find a match in optab,
if no match, look for it in the new path.
1.2, get all candidates in the machine instruction table through opcode.
1.3, check whether the number of arguments is the same, and whether the
class of each argument matches.

2, the instruction encoding process.
The new path encodes an instruction by the following formula:
Binary(Prog) = skeleton|Enc(opcode)|Enc(arg1)|Enc(arg2)|...
where skeleton is the known bits of the machine instruction.

3, the new machine instruction table.
The table was auto-generated at first, but required some manual
adjustments. The table structure is as follows:
type inst struct {
	skeleton uint32
	as       string
	args     instArgs
}
where skeleton is the known bits of the instruction. as is the name of
the Go assembly instruction, not the arm instruction name, and we find
all candidates for a Prog based on this field. args is the instruction
argument type, and the order is consistent with the syntax order of Go
assembly instructions, so that we do not need to adjust the argument
order when encoding.

Updates #44734

Change-Id: I188814700506d76a353f398b76204667d93078d5