cmd/internal/obj/arm64: new path for adding new instructions
This CL adds a new path for adding new instructions, the purpose is to
simplify the effort of adding new instructions.

The new path is based on the assumption that a Go assembly instruction
(a Prog instance) corresponds to only one arm64 machine instruction.
Based on this assumption, matching a Prog to a machine instruction is
much simpler.

This CL preserves the original path, and theoretically based on the above
assumptions, all new instructions can be added with the new path. At the
same time, the old path can be used in the case where one Go assembly
instruction corresponds to multiple machine instructions, if such a
situation exists.

There are three main changes:
1. The process of matching machine instructions for a Prog.
  1.1, For a Prog, first use the existing method to find a match in optab,
       if no match, look for it in the new path.
  1.2, Get all candidates in the machine instruction table through opcode.
  1.3, Check whether the number of arguments is the same, and whether the
       class of each argument matches.

2, The instruction encoding process.
The new path encodes an instruction by the following formula:
Binary(Prog) = skeleton|Enc(opcode)|Enc(arg1)|Enc(arg2)|...
Each argument consists of multiple elements, so
Enc(arg) = Enc(elm1)|Enc(elm2)...
So we changed the encoding of an instruction into the encoding of
argument elements, and this part of work needs to be done manually.
Skeleton is the known bits of the machine instruction.

3, The new machine instruction table.
The table was auto-generated by parsing the public Arm ISA document.
Its structure is as follows:
type inst struct {
	goOp     string // Go opcode mnemonic
	armOp    string // Arm opcode mnemonic
	feature  string // such as "FEAT_LSE", "FEAT_CSSC"
	skeleton uint32 // known bits
	mask     uint32 // mask for disassembly
	args     []arg  // args, in Go order
}
The structure of arg is as follow:
type arg struct {
	aType int         // such as REG, RSP, IMM, COND, PCREL...
	elms []elmType    // the elements that this arg includes
}
Note that the order of args is consistent with the syntax order of Go
assembly instructions, so that we do not need to adjust the argument
order when encoding.

Updates #44734

Change-Id: I188814700506d76a353f398b76204667d93078d5