cmd/compile: package-annotate structs when error would be ambiguous
An attempt at fixing this, perhaps there are more graceful ways
to express it.  Before emitting a "wanted Foo but got Bar" message
for an interface type match failure, check that Foo and Bar are
different.  If they are not, add package paths to non-empty structs,
because that is the cause.

Still figuring out best way to do a test, the important output
is not directly prefixed by a line number, e.g.

./main.go:9:12: cannot use a.S{} ... (wrong type for method M0)
		have M0() (_ struct{f string /* package a */ })
		want M0() (_ struct{f string /* package main */ })
./main.go:15:12: cannot use a.S{} ... (wrong type for method M1)
		have M1() (_ struct{string /* package a */ })
		want M1() (_ struct{string /* package main */ })
./main.go:21:12: cannot use a.S{} ... (wrong type for method M2)
		have M2() (_ struct{f struct{f string} /* package a */ })
		want M2() (_ struct{f struct{f string} /* package main */ })
./main.go:27:12: cannot use a.S{} ... (wrong type for method M3)
		have M3() (_ struct{F struct{f string /* package a */ }})
		want M3() (_ struct{F struct{f string /* package main */ }})

Fixes #54258.

Change-Id: Ifc2b2067d62fe2138996972cdf3b6cb7ca0ed456