cmd/compile: package-annotate structs when error would be ambiguous
An attempt at fixing this, perhaps there are more graceful ways
to express it.  Before emitting a "wanted Foo but got Bar" message
for an interface type match failure, check that Foo and Bar are
different.  If they are not, add package paths to non-empty structs,
because that is the cause.

Still figuring out best way to do a test, but this is the output
of what would be a test if I knew how to make this into a test.

./main.go:13:12: cannot use b.S{} (value of type b.S) as I0 value in variable declaration: b.S does not implement I0 (wrong type for method M0)
		have M0() (struct{f string /* package a/b */ })
		want M0() (struct{f string /* package main */ })
./main.go:19:12: cannot use b.S{} (value of type b.S) as I1 value in variable declaration: b.S does not implement I1 (wrong type for method M1)
		have M1() (struct{string /* package a/b */ })
		want M1() (struct{string /* package main */ })
./main.go:25:12: cannot use b.S{} (value of type b.S) as I2 value in variable declaration: b.S does not implement I2 (wrong type for method M2)
		have M2() (struct{f struct{f string} /* package a/b */ })
		want M2() (struct{f struct{f string} /* package main */ })
./main.go:31:12: cannot use b.S{} (value of type b.S) as I3 value in variable declaration: b.S does not implement I3 (wrong type for method M3)
		have M3() (struct{F struct{f string /* package a/b */ }})
		want M3() (struct{F struct{f string /* package main */ }})
Fixes #54258.

Change-Id: Ifc2b2067d62fe2138996972cdf3b6cb7ca0ed456