cmd/compile: package-annotate structs when error would be ambiguous
An attempt at fixing this, perhaps there are more graceful ways
to express it.  Before emitting a "wanted Foo but got Bar" message
for an interface type match failure, check that Foo and Bar are
different.  If they are not, add package paths to non-empty structs,
because that is the cause.

Added test to test/fixedbugs.

Fixes #54258.

Change-Id: Ifc2b2067d62fe2138996972cdf3b6cb7ca0ed456